Fixed ride estimation

+++ transport_projekat_backend/src/main/java/rs/ac/uns/ftn/transport/dto/ride/RideEstimationDTO.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class RideEstimationDTO {
-    private int estimatedTimeInMinutes;
-    private int estimatedCost;
+    private double estimatedTimeInMinutes;
+    private double estimatedCost;
 }
